requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
configparser==3.5.0
dj-database-url==0.4.2
Django==1.8.4
django-webpack-loader==0.2.4
djangorestframework==3.3.2
djangorestframework-camel-case==0.2.0
enum34==1.1.6
gunicorn==19.7.1
httplib2==0.10.3
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2==1.9.0.post1
pylint==1.7.2
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
